Use Org Display
Move isValidSfdxAuthUrl to module
