+++ packages/sfpowerscripts-cli/src/impl/prepare/PrepareImpl.ts
@@ -3,8 +3,8 @@ import { Org } from "@salesforce/core";
 import { PoolConfig } from "../pool/PoolConfig";
 import { PreparePoolInterface } from "./PreparePoolInterface";
 import PreparePool from "./PreparePool";
-
-
+import OrgDisplayImpl from "@dxatscale/sfpowerscripts.core/lib/sfdxwrappers/OrgDisplayImpl";
+import isValidSfdxAuthUrl from "../pool/prequisitecheck/IsValidSfdxAuthUrl";
 
 
 
@@ -28,7 +28,10 @@ export default class PrepareImpl {
   {
     let poolPreparer:PreparePoolInterface;
 
-    if(!this.hubOrg.getConnection().refreshToken)
+    let orgDisplayImpl = new OrgDisplayImpl(null, this.hubOrg.getUsername());
+    let orgDisplayResult = await orgDisplayImpl.exec(true);
+
+    if(!(orgDisplayResult.sfdxAuthUrl && isValidSfdxAuthUrl(orgDisplayResult.sfdxAuthUrl)))
       throw new  Error(`Pools have to be created using a DevHub authenticated with auth:web or auth:store or auth:accesstoken:store`);
 
     poolPreparer = new PreparePool(this.hubOrg,this.pool);
